+++ packages/embark-transaction-logger/README.md
@@ -1,6 +1,6 @@
-# `embark-console-listener`
+# `embark-transaction-logger`
 
-> Listens to Embark logs and stores them so that they can be printed in Cockpit
+> Listens and logs transactions in a  human readble way
 
 Visit [embark.status.im](https://embark.status.im/) to get started with
 [Embark](https://github.com/embark-framework/embark).
