+++ packages/embark-transaction-logger/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "embark-console-listener",
+  "name": "embark-transaction-logger",
   "version": "4.1.0-beta.5",
   "author": "Iuri Matias <iuri.matias@gmail.com>",
   "contributors": [],
-  "description": "Console listener component for Embark",
-  "homepage": "https://github.com/embark-framework/embark/tree/master/packages/embark-console-listener#readme",
+  "description": "Transaction listener and logger for Embark",
+  "homepage": "https://github.com/embark-framework/embark/tree/master/packages/embark-transaction-logger#readme",
   "bugs": "https://github.com/embark-framework/embark/issues",
   "keywords": [
     "blockchain",
@@ -20,7 +20,7 @@
   ],
   "license": "MIT",
   "repository": {
-    "directory": "packages/embark-console-listener",
+    "directory": "packages/embark-transaction-logger",
     "type": "git",
     "url": "https://github.com/embark-framework/embark.git"
   },
@@ -47,7 +47,12 @@
   },
   "dependencies": {
     "@babel/runtime-corejs2": "7.3.1",
-    "embark-process-logs-api": "^4.1.0-beta.5"
+    "async": "2.6.1",
+    "embark-core": "^4.1.0-beta.5",
+    "embark-i18n": "^4.1.0-beta.3",
+    "embark-utils": "^4.1.0-beta.5",
+    "ethereumjs-tx": "1.3.7",
+    "ethereumjs-util": "6.0.0"
   },
   "devDependencies": {
     "@babel/cli": "7.2.3",
