+++ packages/embark/package.json
@@ -127,6 +127,7 @@
     "embark-storage": "^4.1.0-beta.5",
     "embark-swarm": "^4.1.0-beta.5",
     "embark-test-runner": "^4.1.0-beta.5",
+    "embark-transaction-logger": "^4.1.0-beta.5",
     "embark-transaction-tracker": "^4.1.0-beta.3",
     "embark-ui": "^4.1.0-beta.5",
     "embark-utils": "^4.1.0-beta.5",
