+++ packages/embark/src/lib/core/engine.js
@@ -213,6 +213,7 @@ class Engine {
     this.registerModulePackage('embark-accounts-manager');
     this.registerModulePackage('embark-specialconfigs', {plugins: this.plugins});
     this.registerModulePackage('embark-console-listener');
+    this.registerModulePackage('embark-transaction-logger');
   }
 
   storageComponent() {
